+++ src/cmem/react-flow/configuration/workflow.ts
@@ -1,5 +1,6 @@
 import { EdgeStep } from "./../../../extensions/react-flow/edges/EdgeStep";
 import { NodeDefault } from "./../../../extensions/react-flow/nodes/NodeDefault";
+import { StickyNoteNode } from "./../nodes/StickyNoteNode";
 
 const edgeTypes = {
     default: EdgeStep,
@@ -15,6 +16,7 @@ const nodeTypes = {
     transform: NodeDefault,
     task: NodeDefault,
     workflow: NodeDefault,
+    stickynote: StickyNoteNode,
 };
 
  export {